requirements-old.txt
accelerate==0.23.0
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.12.0
anyio==3.6.2
appdirs==1.4.4
argcomplete==3.1.2
arrow==1.2.3
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.0
beautifulsoup4==4.12.2
bleach==6.0.0
blessed==1.20.0
bokeh==2.4.3
boto==2.49.0
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
clarabel==0.6.0
click==8.1.3
cloudpickle==2.2.1
cmake==3.25.0
colorcet==3.0.1
colorlog==6.7.0
contourpy==1.0.7
crcmod==1.7
croniter==1.3.14
cryptography==41.0.4
cvxpy==1.4.1
cycler==0.11.0
cython==3.0.2
dask==2023.4.1
datashader==0.14.4
datashape==0.5.2
dateutils==0.6.12
deepdiff==6.3.0
detectors==0.1.10
docker-pycreds==0.4.0
docstring-parser==0.15
ecos==2.0.12
faiss-cpu==1.7.4
fastapi==0.88.0
fasteners==0.19
filelock==3.9.0
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
gcs-oauth2-boto-plugin==3.0
gitdb==4.0.10
gitpython==3.1.31
google-apitools==0.5.32
google-auth==2.23.3
google-reauth==0.1.1
greenlet==3.0.0
gsutil==5.26
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
holoviews==1.16.0
httplib2==0.20.4
huggingface-hub==0.15.1
idna==3.4
imageio==2.28.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
inquirer==3.1.3
itsdangerous==2.1.2
jinja2==3.1.2
joblib==1.2.0
jsonargparse==4.21.1
kiwisolver==1.4.4
kornia==0.7.0
lazy-loader==0.2
learn2learn==0.2.0
librosa==0.10.1
lightning==2.0.2
lightning-cloud==0.5.36
lightning-flash==0.8.1.post0
lightning-utilities==0.8.0
lit==15.0.7
littleutils==0.2.2
llvmlite==0.40.0
locket==1.0.0
mako==1.2.4
markdown==3.4.3
markdown-it-py==2.2.0
markupsafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
monotonic==1.6
mpmath==1.2.1
msgpack==1.0.5
multidict==6.0.4
multipledispatch==0.6.0
munkres==1.1.4
networkx==3.0
numba==0.57.0
numpy==1.23.5
oauth2client==4.1.3
ogb==1.3.6
opencv-python==4.7.0.72
optuna==3.4.0
ordered-set==4.1.0
osqp==0.6.3
outdated==0.2.2
packaging==23.1
pandas==2.0.1
panel==0.14.4
param==1.13.0
partd==1.4.0
pathtools==0.1.2
pillow==9.3.0
plotly==5.13.1
pooch==1.7.0
protobuf==3.20.1
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.11.1
pycparser==2.21
pyct==0.5.0
pydantic==1.10.7
pydeprecate==0.3.2
pyjwt==2.7.0
pynndescent==0.5.10
pyopenssl==23.2.0
pyparsing==3.0.9
python-editor==1.0.4
python-multipart==0.0.6
pytorch-fid==0.3.0
pytorch-lightning==1.9.0
pytz==2023.3
pyu2f==0.1.5
pyviz-comms==2.2.1
pywavelets==1.4.1
pyworld==0.3.4
pyyaml==6.0
qdldl==0.1.7.post0
qpth==0.0.16
readchar==4.0.5
requests==2.28.1
retry-decorator==1.1.1
rich==13.3.5
rsa==4.7.2
safetensors==0.3.1
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
scs==3.2.3
seaborn==0.12.2
sentry-sdk==1.23.0
setproctitle==1.3.2
setuptools==59.5.0
smmap==5.0.0
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.6
sqlalchemy==2.0.22
starlette==0.22.0
starsessions==1.3.0
sympy==1.11.1
tenacity==8.2.2
termcolor==2.3.0
tgt==1.4.4
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.9.0
toolz==0.12.0
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typeshed-client==2.3.0
typing-extensions==4.4.0
tzdata==2023.3
umap-learn==0.5.3
urllib3==1.26.13
uvicorn==0.22.0
wandb==0.15.2
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.3
wilds==2.0.0
xarray==2023.4.2
yarl==1.9.2
zipp==3.15.0
timm
learn2learn